Refactors SumOfNumericColumn to work with substitution mechanism
VarSub stuff requires the properties passed in to be delayed processing.
That is, they should be strings or JSON strings until they're needed so
that they can be substituted.

This is exposing a rough edge where our check modeling is bound by the
underlying configuration details. Instead, we should be materializing
a check from config once the config is settled. For example,
a CheckConfig models the YAML/JSON and then CheckConfig#buildCheck
produces a Check without any JSON details. The CheckConfig can be
manipulated as needed until that build executes.