Tweaks the colstats description
The word "value" and the structure of the table made me think that that list of statistical measures needed to be specified in the config. I had to read it a couple times to realize that this is just informational about the things that colstats will output. Suggest a quick tweak to the language to clarify this.