Avoid test failure from envvar missing in some environments
Set JAVA_HOME variable for test scope because in some OSes, it's not set
by default. Notably, Ubuntu doesn't set the envvar in user profiles when
using tooling such as `update-java-alternatives`. This avoids a spurious
test failure becaus of the assumption that this envvar would always be
set. It was, when all dev was done on Macs ;-)