Creates scala-steward GHA workflow
Someone posted on [our submission to the centralized Scala Steward bot][1]
that the bot is offline and may be so indefinitely. Poking around, there's some
movement about reviving it but the direction seems to be to set it up yourself
since GHA makes that very easy except for [the credential management part][3].

For now, we'll just use the repo secret for the token even though tests won't run until
after it's merged. If worse comes to worse, I'll use a personal token until I can figure
out if we have an internal service account I can use.

[1]: https://github.com/scala-steward-org/repos/pull/1030
[2]: https://github.com/scala-steward-org/repos/pull/1029
[3]: https://github.com/scala-steward-org/scala-steward-action#github-token