Adds docs on creating Ivy settings for repo resolution
GitHub Packages currently— and for the foreseeable future— requires credentials to access even public GitHub Packages repositories. Specifying this to `spark-submit` isn't straightforward. It requires creating and passing an `ivysettings` XML file. This patch documents that process.

This also bumps all of the DV versions up to 0.14.1, latest as of this patch.